pyproject.toml
[project]
name = "mask-r-ccn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=22.0.0",
    "albumentations>=1.3.0",
    "black>=23.0.0",
    "bokeh>=3.0.0",
    "detectron2",
    "fastapi>=0.100.0",
    "flake8>=5.0.0",
    "glob2>=0.7",
    "gpustat>=1.1.0",
    "gputil>=1.4.0",
    "hydra-core>=1.3.0",
    "imgaug>=0.4.0",
    "ipdb>=0.13.0",
    "ipykernel>=6.20.0",
    "ipython>=8.10.0",
    "isort>=5.10.0",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "matplotlib>=3.6.0",
    "notebook>=6.5.0",
    "numpy>=1.21.0",
    "nvidia-ml-py3>=7.352.0",
    "onnx>=1.12.0",
    "onnxruntime-gpu>=1.14.0",
    "opencv-python>=4.8.0",
    "pandas>=1.5.0",
    "pillow>=7.1.0,<10.0.0",
    "plotly>=5.10.0",
    "psutil>=5.8.0",
    "pycocotools>=2.0.6",
    "python-dotenv>=0.19.0",
    "python-multipart>=0.0.5",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
    "scikit-image>=0.20.0",
    "scipy>=1.9.0",
    "seaborn>=0.11.0",
    "streamlit>=1.20.0,<1.30.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "torchvision>=0.16.0",
    "tqdm>=4.64.0",
    "uvicorn>=0.20.0",
]

[tool.uv.sources]
detectron2 = { git = "https://github.com/facebookresearch/detectron2.git" }
